split launch into native and html, add configure/html
